[Examples] Upgrade for the new format
I have forgotten to take into account the examples in my lastet PR
to use getMuiTheme.
I though it was a good opportunity to clean our examples.
I have:
- Remove the component folder
- Remove the `_` for the method handlers
- Fix the usage of `getMuiTheme`
- Use the `themeDecorator` to provide the theme.
- Upgraded theme to ES6 classes
I'm convinced that it's should be the only way that we should document
