+++ examples/browserify-gulp-example/package.json
@@ -15,7 +15,7 @@
     "babel-preset-react": "^6.3.13",
     "babelify": "^7.2.0",
     "browser-sync": "^2.11.0",
-    "browserify": "^12.0.1",
+    "browserify": "^13.0.0",
     "gulp": "^3.8.10",
     "gulp-notify": "^2.1.0",
     "gulp-util": "^3.0.1",
