+++ examples/browserify-gulp-example/src/app/Main.jsx
@@ -1,0 +1,79 @@
+/**
+ * In this file, we create a React component
+ * which incorporates components providedby material-ui.
+ */
+
+import React from 'react';
+import RaisedButton from 'material-ui/lib/raised-button';
+import Dialog from 'material-ui/lib/dialog';
+import Colors from 'material-ui/lib/styles/colors';
+import FlatButton from 'material-ui/lib/flat-button';
+import getMuiTheme from 'material-ui/lib/styles/getMuiTheme';
+import themeDecorator from 'material-ui/lib/styles/theme-decorator';
+
+const styles = {
+  container: {
+    textAlign: 'center',
+    paddingTop: 200,
+  },
+};
+
+const muiTheme = getMuiTheme({
+  accent1Color: Colors.deepOrange500,
+});
+
+class Main extends React.Component {
+  constructor(props, context) {
+    super(props, context);
+    this.handleRequestClose = this.handleRequestClose.bind(this);
+    this.handleTouchTap = this.handleTouchTap.bind(this);
+
+    this.state = {
+      open: false,
+    };
+  }
+
+  handleRequestClose() {
+    this.setState({
+      open: false,
+    });
+  }
+
+  handleTouchTap() {
+    this.setState({
+      open: true,
+    });
+  }
+
+  render() {
+    const standardActions = (
+      <FlatButton
+        label="Okey"
+        secondary={true}
+        onTouchTap={this.handleRequestClose}
+      />
+    );
+
+    return (
+      <div style={styles.container}>
+        <Dialog
+          open={this.state.open}
+          title="Super Secret Password"
+          actions={standardActions}
+          onRequestClose={this.handleRequestClose}
+        >
+          1-2-3-4-5
+        </Dialog>
+        <h1>material-ui</h1>
+        <h2>example project</h2>
+        <RaisedButton
+          label="Super Secret Password"
+          primary={true}
+          onTouchTap={this.handleTouchTap}
+        />
+      </div>
+    );
+  }
+}
+
+export default themeDecorator(muiTheme)(Main);
