--- examples/webpack-example/src/app/components/main.jsx
@@ -1,83 +1,0 @@
-/** In this file, we create a React component which incorporates components provided by material-ui */
-
-import React from 'react';
-import RaisedButton from 'material-ui/lib/raised-button';
-import Dialog from 'material-ui/lib/dialog';
-import ThemeManager from 'material-ui/lib/styles/theme-manager';
-import LightRawTheme from 'material-ui/lib/styles/raw-themes/light-raw-theme';
-import Colors from 'material-ui/lib/styles/colors';
-import FlatButton from 'material-ui/lib/flat-button';
-import getMuiTheme from 'material-ui/lib/styles/getMuiTheme';
-
-const containerStyle = {
-  textAlign: 'center',
-  paddingTop: 200,
-};
-
-const Main = React.createClass({
-
-  childContextTypes: {
-    muiTheme: React.PropTypes.object,
-  },
-
-  getInitialState() {
-    return {
-      muiTheme: getMuiTheme(LightRawTheme),
-      open: false,
-    };
-  },
-
-  getChildContext() {
-    return {
-      muiTheme: this.state.muiTheme,
-    };
-  },
-
-  componentWillMount() {
-    let newMuiTheme = ThemeManager.modifyRawThemePalette(this.state.muiTheme, {
-      accent1Color: Colors.deepOrange500,
-    });
-
-    this.setState({muiTheme: newMuiTheme});
-  },
-
-  _handleRequestClose() {
-    this.setState({
-      open: false,
-    });
-  },
-
-  _handleTouchTap() {
-    this.setState({
-      open: true,
-    });
-  },
-
-  render() {
-    const standardActions = (
-      <FlatButton
-        label="Okey"
-        secondary={true}
-        onTouchTap={this._handleRequestClose}
-      />
-    );
-
-    return (
-      <div style={containerStyle}>
-        <Dialog
-          open={this.state.open}
-          title="Super Secret Password"
-          actions={standardActions}
-          onRequestClose={this._handleRequestClose}
-        >
-          1-2-3-4-5
-        </Dialog>
-        <h1>material-ui</h1>
-        <h2>example project</h2>
-        <RaisedButton label="Super Secret Password" primary={true} onTouchTap={this._handleTouchTap} />
-      </div>
-    );
-  },
-});
-
-export default Main;
